feat: refactor Main scene, fix locking screen
